refactor: 캐릭터 init 리팩토링 (usePlayer 공통 hook 생성 및 적용)
