+++ project04/frontend/src/components/content/canvas/maps/player/Man.jsx
@@ -1,30 +1,13 @@
-import React, { useEffect, useMemo, useRef, useState } from "react";
-import { useGLTF, useAnimations } from "@react-three/drei";
-import { SkeletonUtils } from "three-stdlib";
-import { useGraph } from "@react-three/fiber";
+import { usePlayer } from "./hooks/usePlayer";
 
-export function Man({ player, position }) {
-  const playerId = player?.id;
-  const memoizedPosition = useMemo(() => position, []);
-  const playerRef = useRef(null);
-  const { scene, materials, animations } = useGLTF(
-    "/models/CubeGuyCharacter.glb"
+export function Man({ player, position, modelIndex }) {
+  const { playerRef, memoizedPosition, playerId, nodes, materials } = usePlayer(
+    {
+      player,
+      position,
+      modelIndex: modelIndex ?? player.selectedCharacterGlbNameIndex,
+    }
   );
-  const clone = useMemo(() => SkeletonUtils.clone(scene), []);
-  const objectMap = useGraph(clone);
-  const nodes = objectMap.nodes;
-  const [animation, setAnimation] = useState(
-    "CharacterArmature|CharacterArmature|CharacterArmature|Idle"
-  );
-  const { actions } = useAnimations(animations, playerRef);
-
-  useEffect(() => {
-    actions[animation]?.reset().fadeIn(0.5).play();
-
-    return () => {
-      actions[animation]?.fadeOut(0.5);
-    };
-  }, [actions, animation]);
 
   return (
     <group
@@ -55,5 +38,3 @@ export function Man({ player, position }) {
     </group>
   );
 }
-
-useGLTF.preload("/models/CubeGuyCharacter.glb");
